0.16.1
added forced updates
adding REQUIRED to version.info will force the launcher to download the update
added resizing support to homestate so that fix an issue where the applist would be stretched if resized at home
updated initialisestate to support force dupdate
message box uses Arial
fixed initalisestate not checking for or creating resources folder
fixed control + r not redownloading appslist
fixed thirdpartynotices not being downloaded
fixed funky text in the first few seconds of initalisestate
fixed some c4244

+++ SFUI-Whorehouse/Item.hpp
@@ -20,7 +20,7 @@ public:
 
 	sf::Text	name;
 	std::string itemName;
-	int totalHeight;
+	float totalHeight;
 
 	int cardNumber; // used to store depth
 	bool downloaded;
